Moved the public services to a dedicated folder
